+++ src/main/java/inflearn/springadvance/aop/order/aop/AspectV6Advice.java
@@ -1,0 +1,57 @@
+package inflearn.springadvance.aop.order.aop;
+
+import lombok.extern.slf4j.Slf4j;
+import org.aspectj.lang.JoinPoint;
+import org.aspectj.lang.ProceedingJoinPoint;
+import org.aspectj.lang.annotation.After;
+import org.aspectj.lang.annotation.AfterReturning;
+import org.aspectj.lang.annotation.AfterThrowing;
+import org.aspectj.lang.annotation.Around;
+import org.aspectj.lang.annotation.Aspect;
+import org.aspectj.lang.annotation.Before;
+
+@Aspect
+@Slf4j
+public class AspectV6Advice {
+
+    @Around("inflearn.springadvance.aop.order.aop.pointcut.Pointcuts.orderAndService()")
+    public Object doTransaction(ProceedingJoinPoint joinPoint) throws Throwable {
+        try {
+            //@Before
+            log.info("[around][트랜잭션 시작] {}", joinPoint.getSignature());
+            Object result = joinPoint.proceed();
+            //@AfterReturning
+            log.info("[around][트랜잭션 커밋] {}", joinPoint.getSignature());
+            return result;
+        } catch (Exception e) {
+            //@AfterThrowing
+            log.info("[around][트랜잭션 롤백] {}", joinPoint.getSignature());
+            throw e;
+        } finally {
+            //@After
+            log.info("[around][리소스 릴리즈] {}", joinPoint.getSignature());
+        }
+    }
+
+    @Before("inflearn.springadvance.aop.order.aop.pointcut.Pointcuts.orderAndService()")
+    public void doBefore(JoinPoint joinPoint) {
+        log.info("[before] {}", joinPoint.getSignature());
+    }
+
+    @AfterReturning(value = "inflearn.springadvance.aop.order.aop.pointcut.Pointcuts.orderAndService()",
+            returning = "result")
+    public void doReturn(JoinPoint joinPoint, Object result) {
+        log.info("[return] {} return={}", joinPoint.getSignature(), result);
+    }
+
+    @AfterThrowing(value = "inflearn.springadvance.aop.order.aop.pointcut.Pointcuts.orderAndService()",
+            throwing = "ex")
+    public void doThrowing(JoinPoint joinPoint, Exception ex) {
+        log.info("[ex] {} message={}", joinPoint.getSignature(), ex.getMessage());
+    }
+
+    @After(value = "inflearn.springadvance.aop.order.aop.pointcut.Pointcuts.orderAndService()")
+    public void doAfter(JoinPoint joinPoint) {
+        log.info("[after] {}", joinPoint.getSignature());
+    }
+}